Less restrictive attribute loading in GenericDataset (#54)
This commit allows loading attributes from multiple csv files where not
all csv files have data for all basins (when data is split by attributes)
or when not all csv files have data for all attributes (when data is split
by basins).

The use-case is mostly the former: We might have a csv with soil data for
all basins and a csv with climate data for half of the basins. With this
commit, it's possible to use the two csv files (on the set of basins for
which all attributes exist).