Fix GMM & CMAL sampling for multi-timescale runs (#204)
Fixes a bug where uncertainty runs with predict_last_n > 1 would
generate incrorrect predictions due to a mixup of dimensions. This was
discovered in an MTS-LSTM setting, where the hourly branch has
predict_last_n 24. Visually, this resulted in 24-hourly steps in the
predictions.

UMAL and MCD are unaffected because they sample differently.