Proposal to solve the recursive import issue in `load_scaler` (#223)
* Solve the recursive import issue in load_scaler

This commit implements a simple solution for the problem of the recursive imports through `load_scaler`: We move the `load_scaler` function to the data. This works becuase the problem so far was that the evaluation init calls testers, which again have a whole set of dependencies that ultimately end up loading even more stuff until `get_model` is called twice. Thus, by simply moving the function we can now directly import it from nh without having a recursive import.